first version move document modal
